Enable NopEliminatoin inside MO Backend (#6874)
* Refactor NopElimination; Execute NopEliminatoin inside MOC Backend

* Add missing header

* Refactor tests; use weak_ptr to avoid excess consumers for node

* Add test to check number of shared_ptr usages for graph nodes during Manager execution

* Remove AlgebraicSimplification forewer
