+++ inference-engine/src/transformations/src/transformations/common_optimizations/nop_elimination.cpp
@@ -5,11 +5,10 @@
 #include "itt.hpp"
 #include <functional>
 #include <memory>
-#include <typeindex>
-#include <typeinfo>
-#include <unordered_map>
+#include <numeric>
 
 #include <ngraph/opsets/opset3.hpp>
+#include <ngraph/opsets/opset8.hpp>
 #include <ngraph/util.hpp>
 #include <ngraph/log.hpp>
 #include <transformations/common_optimizations/nop_elimination.hpp>
@@ -18,9 +17,60 @@
 using namespace std;
 using namespace ngraph;
 
-#define TI(x) x::type_info
+//`simplify_gather`, optimizes gather if Gather is gathering the
+// whole input tensor
+static bool simplify_gather(std::shared_ptr<Node> node) {
+    if (auto gather = as_type_ptr<opset3::Gather>(node)) {
+        // check if we are gathering the whole input
+        auto data = gather->input_value(0);
+        auto indices = gather->input_value(1);
 
-NGRAPH_RTTI_DEFINITION(ngraph::pass::NopElimination, "NopElimination", 0);
+        // we need to know data and indices shape to infer if gather is Nop
+        if (data.get_partial_shape().is_dynamic() || indices.get_partial_shape().is_dynamic()) {
+            return false;
+        }
+        // if rank of data and gather output dont match, we will skip
+        if (data.get_shape().size() != node->get_shape().size()) {
+            return false;
+        }
+
+        auto axis = gather->get_axis();
+        if (axis == opset3::Gather::AXIS_NOT_SET_VALUE) {
+            NGRAPH_DEBUG << "axis value not set";
+            return false;
+        }
+
+        // case_1 : if the input tensor is of shape (4, 1, 4)
+        // and axis = 1, then the gather would be simply
+        // gathering the whole input tensor, so we can optimize this
+        // op has Nop
+
+        if (data.get_shape()[axis] == 1 && data.get_shape() == node->get_shape()) {
+            return replace_output_update_name(gather->output(0), gather->input_value(0));
+        }
+
+        // case_2 : if the input tensor is of shape (4, 3, 4)
+        // we need to check the contents of indices, if indices
+        // is 1D tensor of value {0, 1, 2}, we can optimize this
+        // op has Nop
+
+        // check if the indices is constant
+        auto constant_indices =
+                as_type_ptr<opset3::Constant>(gather->input_value(1).get_node_shared_ptr());
+        if (!constant_indices) {
+            return false;
+        } else {
+            // if ref_inidices == indices, we are capturing the
+            // entire input tensor
+            std::vector<int64_t> ref_indices(data.get_shape()[axis], 0);
+            std::iota(ref_indices.begin(), ref_indices.end(), 0);
+            if (ref_indices == constant_indices->cast_vector<int64_t>()) {
+                return replace_output_update_name(gather->output(0), gather->input_value(0));
+            }
+        }
+    }
+    return false;
+}
 
 static bool eliminate_nop(const std::shared_ptr<Node>& node) {
     // skip if shapes are dynamic
@@ -31,34 +81,6 @@ static bool eliminate_nop(const std::shared_ptr<Node>& node) {
 
     if (node->get_input_shape(0) == node->get_output_shape(0)) {
         return replace_output_update_name(node->output(0), node->input_value(0));
-    }
-    return false;
-}
-
-static bool eliminate_convert(const std::shared_ptr<Node>& node) {
-    bool is_out_type_agnostic = false;
-    static const std::set<NodeTypeInfo> type_agnostic{TI(opset3::NonZero)};
-    if (node->output(0).get_target_inputs().size() == 1) {
-        Input<Node> out = *node->output(0).get_target_inputs().begin();
-        is_out_type_agnostic = type_agnostic.count(out.get_node()->get_type_info()) == 1;
-    }
-    auto convert = as_type_ptr<opset3::Convert>(node);
-    auto input = convert->input_value(0);
-    if (convert->get_convert_element_type() == input.get_element_type() || is_out_type_agnostic) {
-        if (is_out_type_agnostic && is_type<opset3::Convert>(input.get_node())) {
-            input = input.get_node()->input_value(0);
-        }
-        return replace_output_update_name(node->output(0), input);
-    }
-    return false;
-}
-
-static bool eliminate_concat(const std::shared_ptr<Node>& node) {
-    auto node_input = node->input_value(0);
-
-    // remove concat with single input
-    if (node->get_input_size() == 1) {
-        return replace_output_update_name(node->output(0), node_input);
     }
     return false;
 }
@@ -336,20 +358,168 @@ NAME() { \
 }; \
 NGRAPH_RTTI_DEFINITION(NAME, STR(NAME), 0);
 
-SIMPLE_MATCHER_PASS_DEFINITION(EliminatePad, opset3::Pad, eliminate_nop);
-SIMPLE_MATCHER_PASS_DEFINITION(EliminateConvert, opset3::Convert, eliminate_convert);
 SIMPLE_MATCHER_PASS_DEFINITION(EliminateReshape, opset3::Reshape, eliminate_reshape_v1);
-SIMPLE_MATCHER_PASS_DEFINITION(EliminateConcat, opset3::Concat, eliminate_concat);
 SIMPLE_MATCHER_PASS_DEFINITION(EliminateSqueeze, opset3::Squeeze, eliminate_squeeze);
 SIMPLE_MATCHER_PASS_DEFINITION(EliminateUnsqueeze, opset3::Unsqueeze, eliminate_unsqueeze);
 SIMPLE_MATCHER_PASS_DEFINITION(EliminateBroadcast, op::v1::Broadcast, eliminate_nop);
+SIMPLE_MATCHER_PASS_DEFINITION(EliminateGather, opset3::Gather, simplify_gather);
 
-ngraph::pass::NopElimination::NopElimination() {
+
+NGRAPH_RTTI_DEFINITION(pass::EliminatePad, "EliminatePad", 0);
+
+pass::EliminatePad::EliminatePad() {
+    MATCHER_SCOPE(EliminatePad);
+    auto pad_node_pattern = pattern::wrap_type<opset8::Pad>();
+
+    matcher_pass_callback callback = [=](pattern::Matcher& m) {
+        auto pad = m.get_match_root();
+
+        auto pad_begin_const = ngraph::get_constant_from_source(pad->input_value(1));
+        auto pad_end_const = ngraph::get_constant_from_source(pad->input_value(2));
+
+        if (!pad_begin_const || !pad_end_const) {
+            return false;
+        }
+
+        const auto pad_begin_value = pad_begin_const->cast_vector<int64_t>();
+        const auto pad_end_value = pad_end_const->cast_vector<int64_t>();
+
+        if (std::any_of(pad_begin_value.begin(), pad_begin_value.end(), [](int64_t value) { return value != 0; }) ||
+            std::any_of(pad_end_value.begin(), pad_end_value.end(), [](int64_t value) { return value != 0; })) {
+            return false;
+        }
+
+        return replace_output_update_name(pad->output(0), pad->input_value(0));
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(pad_node_pattern, matcher_name);
+    this->register_matcher(m, callback);
+}
+
+NGRAPH_RTTI_DEFINITION(pass::EliminateConvert, "EliminateConvert", 0);
+
+pass::EliminateConvert::EliminateConvert() {
+    MATCHER_SCOPE(EliminateConvert);
+    auto convert_pattern = pattern::wrap_type<opset8::Convert>();
+
+    matcher_pass_callback callback = [](pattern::Matcher& m) {
+        auto convert = std::dynamic_pointer_cast<opset8::Convert>(m.get_match_root());
+        if (!convert) {
+            return false;
+        }
+        if (convert->get_input_element_type(0) == convert->get_element_type()) {
+            return replace_output_update_name(convert->output(0), convert->input_value(0));
+        }
+        return false;
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(convert_pattern, matcher_name);
+    this->register_matcher(m, callback);
+}
+
+NGRAPH_RTTI_DEFINITION(pass::EliminateConvertNonZero, "EliminateConvertNonZero", 0);
+
+pass::EliminateConvertNonZero::EliminateConvertNonZero() {
+    MATCHER_SCOPE(EliminateConvertNonZero);
+    auto convert_pattern = pattern::wrap_type<opset8::Convert>(pattern::consumers_count(1));
+    auto non_zero = pattern::wrap_type<opset8::NonZero>({convert_pattern});
+
+    matcher_pass_callback callback = [=](pattern::Matcher& m) {
+        const auto & pattern_map = m.get_pattern_map();
+        auto convert = pattern_map.at(convert_pattern);
+        // remove convert
+        convert->output(0).replace(convert->input_value(0));
+        // to make this elimination recursive we register NonZero as a node which will be used to repeat matching
+        register_new_node(m.get_match_root());
+        return true;
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(non_zero, matcher_name);
+    this->register_matcher(m, callback);
+}
+
+NGRAPH_RTTI_DEFINITION(pass::EliminateConcat, "EliminateConcat", 0);
+
+pass::EliminateConcat::EliminateConcat() {
+    MATCHER_SCOPE(EliminateConcat);
+    auto convert_pattern = pattern::wrap_type<opset8::Concat>();
+
+    matcher_pass_callback callback = [](pattern::Matcher& m) {
+        auto concat = m.get_match_root();
+        if (concat->inputs().size() == 1) {
+            return replace_output_update_name(concat->output(0), concat->input_value(0));
+        }
+        return false;
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(convert_pattern, matcher_name);
+    this->register_matcher(m, callback);
+}
+
+NGRAPH_RTTI_DEFINITION(pass::EliminateSplit, "EliminateSplit", 0);
+
+pass::EliminateSplit::EliminateSplit() {
+    MATCHER_SCOPE(EliminateConcat);
+    auto convert_pattern = pattern::wrap_type<opset8::Split>();
+
+    matcher_pass_callback callback = [](pattern::Matcher& m) {
+        auto split = std::dynamic_pointer_cast<opset8::Split>(m.get_match_root());
+        if (!split || split->get_num_splits() != 1) {
+            return false;
+        }
+        return replace_output_update_name(split->output(0), split->input_value(0));
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(convert_pattern, matcher_name);
+    this->register_matcher(m, callback);
+}
+
+NGRAPH_RTTI_DEFINITION(pass::EliminateTranspose, "EliminateTranspose", 0);
+
+pass::EliminateTranspose::EliminateTranspose() {
+    MATCHER_SCOPE(EliminateTranspose);
+    auto order = pattern::wrap_type<opset8::Constant>();
+    auto transpose_pattern = pattern::wrap_type<opset8::Transpose>({pattern::any_input(), order});
+
+    matcher_pass_callback callback = [=](pattern::Matcher& m) {
+        const auto & pattern_map = m.get_pattern_map();
+        auto order_const = std::dynamic_pointer_cast<opset8::Constant>(pattern_map.at(order));
+        if (!order_const) {
+            return false;
+        }
+
+        const auto & order_values = order_const->cast_vector<int64_t>();
+        vector<int64_t> ref_values(order_values.size());
+        std::iota(ref_values.begin(), ref_values.end(), 0);
+        if (order_values != ref_values) {
+            return false;
+        }
+
+        auto transpose = m.get_match_root();
+        return replace_output_update_name(transpose->output(0), transpose->input_value(0));
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(transpose_pattern, matcher_name);
+    this->register_matcher(m, callback);
+}
+
+NGRAPH_RTTI_DEFINITION(ngraph::pass::NopElimination, "NopElimination", 0);
+
+ngraph::pass::NopElimination::NopElimination(bool use_shape_for_elimination) {
+    // shape-agnostic transformations
     add_matcher<EliminatePad>();
     add_matcher<EliminateConvert>();
-    add_matcher<EliminateReshape>();
+    add_matcher<EliminateConvertNonZero>();
     add_matcher<EliminateConcat>();
-    add_matcher<EliminateSqueeze>();
-    add_matcher<EliminateUnsqueeze>();
-    add_matcher<EliminateBroadcast>();
+    add_matcher<EliminateSplit>();
+    add_matcher<EliminateTranspose>();
+
+    // shape-dependent transformations
+    if (use_shape_for_elimination) {
+        add_matcher<EliminateReshape>();
+        add_matcher<EliminateSqueeze>();
+        add_matcher<EliminateUnsqueeze>();
+        add_matcher<EliminateBroadcast>();
+        add_matcher<EliminateGather>();
+    }
 }
