+++ inference-engine/tests/functional/inference_engine/transformations/transpose_to_reshape_test.cpp
@@ -17,7 +17,6 @@
 #include <ngraph/pass/constant_folding.hpp>
 #include <transformations/utils/utils.hpp>
 #include <transformations/init_node_info.hpp>
-#include <transformations/common_optimizations/algebraic_simplification.hpp>
 #include <ngraph/pass/visualize_tree.hpp>
 #include <transformations/common_optimizations/transpose_sinking.hpp>
 #include <transformations/common_optimizations/transpose_to_reshape.hpp>
@@ -25,6 +24,8 @@
 #include "common_test_utils/ngraph_test_utils.hpp"
 
 using namespace testing;
+using namespace ngraph;
+using namespace std;
 
 using InputShape = ngraph::PartialShape;
 using TransposeOrder = std::vector<int64_t>;
@@ -140,3 +141,93 @@ INSTANTIATE_TEST_SUITE_P(ReshapeWithGather, TransposeToReshapeTests,
 #undef SAME_FUNCTION
 #undef EMPTY_FUNCTION
 #undef SHAPE_OF_GATHER
+
+TEST(TransformationTests, replace_transpose_with_reshape) {
+    auto check_usecase = [](const PartialShape& shape,
+                            const std::vector<int64_t>& perm_val,
+                            bool i32,
+                            bool multiout,
+                            size_t num) {
+        static size_t id = 0;
+        auto casename = string("usecase #") + to_string(++id);
+
+        shared_ptr<Node> perm;
+        if (i32) {
+            std::vector<int32_t> perm_val_i32(perm_val.begin(), perm_val.end());
+            perm =
+                    op::Constant::create<int32_t>(element::i32, Shape{perm_val.size()}, perm_val_i32);
+        } else {
+            perm = op::Constant::create<int64_t>(element::i64, Shape{perm_val.size()}, perm_val);
+        }
+        auto param = make_shared<op::Parameter>(element::f32, shape);
+        shared_ptr<Node> A1;
+        if (multiout) {
+            shared_ptr<Node> k;
+            auto last_dim = shape.rank().get_length() - 1;
+            if (shape[last_dim].is_dynamic()) {
+                k = make_shared<op::v1::Gather>(make_shared<op::ShapeOf>(param),
+                                                op::Constant::create(element::i64, {}, {last_dim}),
+                                                op::Constant::create(element::i64, {}, {0}));
+            } else {
+                k = make_shared<op::Constant>(element::i64, Shape{}, std::vector<int64_t>{shape[last_dim].get_length()});
+            }
+            A1 = make_shared<op::v1::TopK>(param, k, last_dim,
+                                           op::v1::TopK::Mode::MAX, op::v1::TopK::SortType::NONE);
+        } else {
+            A1 = make_shared<op::v0::Abs>(param);
+        }
+        auto transpose = make_shared<op::v1::Transpose>((multiout ? A1->output(0) : A1), perm);
+        auto transpose1 = make_shared<op::v0::Abs>(transpose);
+        auto baseline_f = make_shared<Function>(transpose1, ParameterVector{param});
+        auto optimized_f = clone_function(*baseline_f);
+
+        pass::Manager pass_manager;
+        pass_manager.register_pass<pass::Validate>();
+        pass_manager.register_pass<pass::TransposeToReshape>();
+        pass_manager.run_passes(optimized_f);
+
+        auto ps = baseline_f->get_results()[0]->get_output_partial_shape(0);
+        auto ps_r = optimized_f->get_results()[0]->get_output_partial_shape(0);
+        EXPECT_TRUE(ps.rank().is_static() && ps_r.rank().is_static()) << casename;
+        ASSERT_EQ(ps.rank().get_length(), ps_r.rank().get_length()) << casename;
+
+        ASSERT_EQ(count_ops_of_type<op::v1::Transpose>(baseline_f), 1);
+        ASSERT_EQ(count_ops_of_type<op::v1::Reshape>(baseline_f), 0);
+        ASSERT_EQ(count_ops_of_type<op::v1::Transpose>(optimized_f), num);
+        ASSERT_EQ(count_ops_of_type<op::v1::Reshape>(optimized_f), (num ? 0 : 1));
+    };
+
+    for (auto& i32 : {true, false})
+        for (auto& multiout : {true, false}) {
+            check_usecase(Shape{1, 3}, vector<int64_t>{1, 0}, i32, multiout, 0);
+            check_usecase(Shape{2, 3, 1}, vector<int64_t>{2, 0, 1}, i32, multiout, 0);
+            check_usecase(Shape{10, 20, 1, 1}, vector<int64_t>{0, 2, 3, 1}, i32, multiout, 0);
+            check_usecase(Shape{10, 1, 1, 20}, vector<int64_t>{0, 3, 1, 2}, i32, multiout, 0);
+            check_usecase(Shape{10, 20, 1, 2}, vector<int64_t>{0, 2, 1, 3}, i32, multiout, 0);
+            check_usecase(Shape{10, 1, 1, 1, 20}, vector<int64_t>{0, 4, 1, 2, 3}, i32, multiout, 0);
+            check_usecase(Shape{10, 20, 1, 1, 1}, vector<int64_t>{0, 2, 3, 4, 1}, i32, multiout, 0);
+            check_usecase(Shape{10, 1, 1, 1, 1}, vector<int64_t>{1, 4, 2, 3, 0}, i32, multiout, 0);
+            check_usecase(Shape{10, 1, 1, 1, 1}, vector<int64_t>{4, 2, 0, 1, 3}, i32, multiout, 0);
+            check_usecase(Shape{10, 20, 1, 2}, vector<int64_t>{0, 2, 3, 1}, i32, multiout, 1);
+            check_usecase(Shape{10, 20, 1, 2}, vector<int64_t>{0, 3, 1, 2}, i32, multiout, 1);
+            check_usecase(Shape{10, 20}, vector<int64_t>{1, 0}, i32, multiout, 1);
+
+            check_usecase(PartialShape{Dimension::dynamic(), 20, 1, 1},
+                          vector<int64_t>{
+                                  0, 2, 3, 1,
+                          },
+                          i32,
+                          multiout,
+                          0);
+            check_usecase(PartialShape{Dimension::dynamic(), Dimension::dynamic(), 20, 1, 1},
+                          vector<int64_t>{0, 1, 3, 2, 4},
+                          i32,
+                          multiout,
+                          0);
+            check_usecase(PartialShape{Dimension::dynamic(), Dimension::dynamic(), 20, 1, 1},
+                          vector<int64_t>{0, 2, 1, 4, 3},
+                          i32,
+                          multiout,
+                          1);
+        }
+}
